Added the necessary data struct holder for entities
