+++ src/containers/entity.rs
@@ -1,3 +1,9 @@
+mod mapitem_data;
+mod npc_data;
 mod player_data;
+mod shared_data;
 
+pub use mapitem_data::*;
+pub use npc_data::*;
 pub use player_data::*;
+pub use shared_data::*;
