+++ src/gameloop/handledata/mapper.rs
@@ -1,6 +1,6 @@
 use super::routes;
 use crate::{
-    containers::{Storage, World},
+    containers::{GlobalKey, Storage, World},
     gametypes::*,
     socket::*,
 };
