+++ src/maps/map.rs
@@ -1,5 +1,5 @@
 use crate::{
-    containers::{HashSet, IndexSet, Storage},
+    containers::{GlobalKey, HashSet, IndexSet, Storage},
     gametypes::*,
     time_ext::MyInstant,
 };
