+++ src/npcs/npc.rs
@@ -2,63 +2,6 @@ use crate::{containers::*, gametypes::*, tasks::*, time_ext::MyInstant};
 use educe::Educe;
 use std::collections::VecDeque;
 
-#[derive(Debug, Copy, Clone, PartialEq, Eq, Default)]
-pub struct NpcIndex(pub u64);
-
-#[derive(Educe, Debug, Copy, Clone, PartialEq, Eq)]
-#[educe(Default)]
-pub struct NpcTimer {
-    #[educe(Default = MyInstant::now())]
-    pub despawntimer: MyInstant,
-    #[educe(Default = MyInstant::now())]
-    pub spawntimer: MyInstant,
-}
-
-#[derive(Educe, Debug, Copy, Clone, PartialEq, Eq)]
-#[educe(Default)]
-pub struct NpcAITimer(#[educe(Default = MyInstant::now())] pub MyInstant); //for rebuilding the a* paths
-
-#[derive(Educe, Debug, Copy, Clone, PartialEq, Eq)]
-#[educe(Default)]
-pub struct NpcPathTimer {
-    #[educe(Default = MyInstant::now())]
-    pub timer: MyInstant,
-    pub tries: usize,
-    //when failing to move due to blocks in movement.
-    pub fails: usize,
-} //for rebuilding the a* paths
-
-#[derive(Educe, Debug, Copy, Clone, PartialEq, Eq)]
-#[educe(Default)]
-pub struct NpcDespawns(#[educe(Default = false)] pub bool);
-
-#[derive(Educe, Debug, Copy, Clone, PartialEq, Eq)]
-#[educe(Default)]
-pub struct NpcMoving(#[educe(Default = false)] pub bool);
-
-#[derive(Educe, Debug, Copy, Clone, PartialEq, Eq)]
-#[educe(Default)]
-pub struct NpcRetreating(#[educe(Default = false)] pub bool);
-
-#[derive(Educe, Debug, Copy, Clone, PartialEq, Eq)]
-#[educe(Default)]
-pub struct NpcWalkToSpawn(#[educe(Default = false)] pub bool);
-
-#[derive(Educe, Debug, Clone, PartialEq, Eq)]
-#[educe(Default)]
-//offset for special things so the npc wont to events based on this spawn time.
-pub struct NpcHitBy(#[educe(Default = Vec::new())] pub Vec<(u32, u64, u64)>);
-
-#[derive(Educe, Debug, Clone, PartialEq, Eq)]
-#[educe(Default)]
-pub struct NpcMoves(#[educe(Default = VecDeque::new())] pub VecDeque<(Position, u8)>);
-
-#[derive(Default, Debug, Copy, Clone, PartialEq, Eq)]
-pub struct NpcSpawnedZone(pub Option<usize>);
-
-#[derive(Default, Debug, Copy, Clone, PartialEq, Eq)]
-pub struct NpcMovePos(pub Option<Position>);
-
 #[inline(always)]
 pub fn is_npc_same(from_entity: GlobalKey, to_entity: GlobalKey) -> bool {
     from_entity == to_entity
