+++ src/players/player.rs
@@ -4,182 +4,6 @@ use mmap_bytey::{MByteBufferRead, MByteBufferWrite};
 use serde::{Deserialize, Serialize};
 use std::sync::{Arc, Mutex};
 
-#[derive(Clone, Debug, Bundle)]
-pub struct Socket {
-    // IP address
-    pub addr: Arc<String>,
-    // Socket ID
-    pub id: usize,
-    // Packet Buffer
-    pub buffer: Arc<Mutex<ByteBuffer>>,
-}
-
-impl Socket {
-    #[inline(always)]
-    pub fn new(id: usize, addr: String) -> Result<Self> {
-        Ok(Self {
-            id,
-            addr: Arc::new(addr),
-            buffer: Arc::new(Mutex::new(ByteBuffer::with_capacity(8192)?)),
-        })
-    }
-}
-
-#[derive(Copy, Clone, Debug, Educe)]
-#[educe(Default)]
-pub struct PlayerItemTimer {
-    #[educe(Default = MyInstant::now())]
-    pub itemtimer: MyInstant,
-}
-
-#[derive(Copy, Clone, Debug, Educe)]
-#[educe(Default)]
-pub struct PlayerMapTimer {
-    #[educe(Default = MyInstant::now())]
-    pub mapitemtimer: MyInstant,
-}
-
-#[derive(
-    PartialEq,
-    Eq,
-    Clone,
-    Debug,
-    Educe,
-    Deserialize,
-    Serialize,
-    ByteBufferRead,
-    ByteBufferWrite,
-    MByteBufferRead,
-    MByteBufferWrite,
-)]
-#[educe(Default)]
-pub struct Inventory {
-    #[educe(Default = (0..MAX_INV).map(|_| Item::default()).collect())]
-    pub items: Vec<Item>,
-}
-
-#[derive(
-    PartialEq,
-    Eq,
-    Clone,
-    Debug,
-    Educe,
-    Deserialize,
-    Serialize,
-    ByteBufferRead,
-    ByteBufferWrite,
-    MByteBufferRead,
-    MByteBufferWrite,
-)]
-#[educe(Default)]
-pub struct TradeItem {
-    #[educe(Default = (0..MAX_TRADE_SLOT).map(|_| Item::default()).collect())]
-    pub items: Vec<Item>,
-}
-
-#[derive(Copy, Clone, Debug, Default)]
-pub struct TradeMoney {
-    pub vals: u64,
-}
-
-#[derive(
-    PartialEq,
-    Eq,
-    Copy,
-    Clone,
-    Debug,
-    Default,
-    Deserialize,
-    Serialize,
-    ByteBufferRead,
-    ByteBufferWrite,
-    MByteBufferRead,
-    MByteBufferWrite,
-)]
-pub enum TradeStatus {
-    #[default]
-    None,
-    Accepted,
-    Submitted,
-}
-
-#[derive(Copy, Clone, Debug, Educe)]
-#[educe(Default)]
-pub struct TradeRequestEntity {
-    #[educe(Default = None)]
-    pub entity: Option<GlobalKey>,
-    #[educe(Default = MyInstant::now())]
-    pub requesttimer: MyInstant,
-}
-
-#[derive(
-    PartialEq,
-    Eq,
-    Clone,
-    Debug,
-    Educe,
-    Deserialize,
-    Serialize,
-    ByteBufferRead,
-    ByteBufferWrite,
-    MByteBufferRead,
-    MByteBufferWrite,
-)]
-#[educe(Default)]
-pub struct PlayerStorage {
-    #[educe(Default = (0..MAX_STORAGE).map(|_| Item::default()).collect())]
-    pub items: Vec<Item>,
-}
-
-#[derive(
-    PartialEq,
-    Eq,
-    Clone,
-    Debug,
-    Educe,
-    Deserialize,
-    Serialize,
-    ByteBufferRead,
-    ByteBufferWrite,
-    MByteBufferRead,
-    MByteBufferWrite,
-)]
-#[educe(Default)]
-pub struct Equipment {
-    #[educe(Default = (0..MAX_EQPT).map(|_| Item::default()).collect())]
-    pub items: Vec<Item>,
-}
-
-#[derive(Copy, Clone, Debug, Default)]
-pub struct Sprite {
-    pub id: u16,
-}
-
-#[derive(Copy, Clone, Debug, Default)]
-pub struct Money {
-    pub vals: u64,
-}
-
-#[derive(Clone, Debug, Default)]
-pub struct ReloginCode {
-    pub code: String,
-}
-
-#[derive(Clone, Debug, Default)]
-pub struct LoginHandShake {
-    pub handshake: String,
-}
-
-#[derive(Copy, Clone, Debug, Default)]
-pub struct Player {
-    pub levelexp: u64,
-    pub useditemid: u32,
-    pub resetcount: i16,
-    pub pvpon: bool,
-    pub pk: bool,
-    pub movesavecount: u16,
-}
-
 pub fn is_player_online(world: &mut World, entity: GlobalKey) -> Result<bool> {
     Ok(
         *world.get::<&WorldEntityType>(entity.0)? == WorldEntityType::Player
