+++ src/tasks/mapswitchtasks.rs
@@ -1,7 +1,7 @@
 use std::cmp::max;
 
 use crate::{
-    containers::{HashSet, Storage, World},
+    containers::{GlobalKey, HashSet, Storage, World},
     gametypes::*,
     maps::*,
     players::*,
